% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DelimitedFile.R
\name{DelimitedFile}
\alias{DelimitedFile}
\title{DelimitedFile class
Class provides convenience for reading and returning basic statistics of text separated rows of a delimited data file.}
\description{
The class uses the data.table::fread to read text separated rows of a file. The class
provides public methods for returning per variable names/types,  descriptive statistics. Note that
the filename extension (such as .csv) is irrelevant. Separator detection is entirely
driven by the file contents.  Users are able to set their own separation character with the `sep` parameter.
}
\author{
Rick Dean
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{file_path}}{A string that defines the full file path to a csv formatted file.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{DelimitedFile$new()}}
\item \href{#method-read}{\code{DelimitedFile$read()}}
\item \href{#method-names}{\code{DelimitedFile$names()}}
\item \href{#method-min_max}{\code{DelimitedFile$min_max()}}
\item \href{#method-numeric_summary}{\code{DelimitedFile$numeric_summary()}}
\item \href{#method-clone}{\code{DelimitedFile$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Defines the classes' public fields
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DelimitedFile$new(file_path = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{file_path}}{A string that defines the full file path to a csv formatted file.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-read"></a>}}
\if{latex}{\out{\hypertarget{method-read}{}}}
\subsection{Method \code{read()}}{
Read the csv file
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DelimitedFile$read(
  sep = "auto",
  header = "auto",
  na_strings = "NA",
  stringsAsFactors = FALSE,
  skip = 0,
  select = NULL,
  drop = NULL,
  col_names,
  check_names = FALSE,
  fill = FALSE,
  logical01 = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sep}}{The separator between columns. Defaults to the character in the set [, `tab` | ; :].
Use NULL or "" to specify no separator; i.e. each line a single character column.}

\item{\code{header}}{Does the first data line contain column names?}

\item{\code{na_strings}}{A character vector of strings which are interpreted as `NA` values.}

\item{\code{stringsAsFactors}}{If TRUE, convert all character columns to factors}

\item{\code{skip}}{If 0 start on the first line and from there finds the first row with a consistent number of columns.}

\item{\code{select}}{A vector of column names or numbers to keep, drop the rest.}

\item{\code{drop}}{A vector of column names or numbers to drop.}

\item{\code{col_names}}{A vector of optional names for the variables(columns).}

\item{\code{check_names}}{If TRUE then the names of the variables are checked to ensure that they are syntactically valid names.}

\item{\code{fill}}{If TRUE then in case the rows have unequal length, blank fields are implicitly filled.}

\item{\code{logical01}}{If TRUE a column containing only 0's and 1's will be read as logical, otherwise as integer.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-names"></a>}}
\if{latex}{\out{\hypertarget{method-names}{}}}
\subsection{Method \code{names()}}{
Return the column names of the data file.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DelimitedFile$names()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-min_max"></a>}}
\if{latex}{\out{\hypertarget{method-min_max}{}}}
\subsection{Method \code{min_max()}}{
Return the minimum and maximum values for a specific variable(column).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DelimitedFile$min_max(columns)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{columns}}{The column names of interest for min/max values}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-numeric_summary"></a>}}
\if{latex}{\out{\hypertarget{method-numeric_summary}{}}}
\subsection{Method \code{numeric_summary()}}{
Return the numeric summary of the data file.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DelimitedFile$numeric_summary(columns)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{columns}}{The column names of interest to summarize}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DelimitedFile$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
